ENG-12929 try a different testing approach: let PullSocketImporter connect to us, and use those sockets to determine if importers are up.
