+++ src/frontend/org/voltdb/importclient/socket/ServerSocketImporter.java
@@ -77,12 +77,10 @@ public class ServerSocketImporter extends AbstractImporter {
         for (ClientConnectionHandler client : m_clients) {
             client.stopClient();
         }
-        info(null, "Stopped");
     }
 
     private void startListening()
     {
-        info(null, "Running");
         try {
             while (shouldRun()) {
                 Socket clientSocket = m_config.getServerSocket().accept();
@@ -98,7 +96,6 @@ public class ServerSocketImporter extends AbstractImporter {
                 warn(e, "Unexpected error accepting client connections for " + getName() + " on port " + m_config.getPort());
             }
         }
-        info(null, "No longer running");
     }
 
     //This is ClientConnection handler to read and dispatch data to stored procedure.
