+++ src/frontend/org/voltdb/importer/ImportManager.java
@@ -288,10 +288,7 @@ public class ImportManager implements ChannelChangeCallback {
         if (m_processor.get() == null) {
             return;
         }
-        //if (m_serverStarted) {
-            m_processor.get().shutdown();
-        //}
-        //Unset until it gets started.
+        m_processor.get().shutdown();
         m_processor.set(null);
     }
