+++ tests/frontend/org/voltdb/TestImportSuite.java
@@ -43,7 +43,6 @@ import org.voltdb.client.Client;
 import org.voltdb.client.ClientImpl;
 import org.voltdb.client.ClientResponse;
 import org.voltdb.compiler.VoltProjectBuilder;
-import org.voltdb.importer.SocketImporterConnector;
 import org.voltdb.regressionsuites.LocalCluster;
 import org.voltdb.regressionsuites.MultiConfigSuiteBuilder;
 import org.voltdb.regressionsuites.RegressionSuite;
@@ -67,7 +66,6 @@ public class TestImportSuite extends RegressionSuite {
 
     private Boolean m_socketHandlerInitialized = false;
     private Client m_client;
-    private Client m_adminClient;
     private List<DataPusher> m_dataPushers = new ArrayList<>();
     private CountDownLatch m_dataAwaiter = null;
 
@@ -81,7 +79,6 @@ public class TestImportSuite extends RegressionSuite {
         super.setUp();
 
         m_client = getClient();
-        m_adminClient = getAdminClient();
         while (!((ClientImpl) m_client).isHashinatorInitialized()) {
             Thread.sleep(1000);
             System.out.println("Waiting for hashinator to be initialized...");
@@ -105,9 +102,48 @@ public class TestImportSuite extends RegressionSuite {
     @Override
     public void tearDown() throws Exception {
         m_dataPushers.clear();
-        releaseClient(m_client);
-        releaseClient(m_adminClient);
+        m_client.close();
         super.tearDown();
+    }
+
+    /** Allows the test thread to check if the importer is alive by writing data to it. */
+    static class ImporterConnector {
+        private final String m_server;
+        private final int m_port;
+        private final char m_separator;
+        private int m_counter = 0;
+
+        public ImporterConnector(String server, int port, char separator) {
+            m_separator = separator;
+            m_server = server;
+            m_port = port;
+        }
+
+        /** Tries to write data to the importer, but may fail. */
+        public void tryPush(int maxAttempts) throws IOException {
+            int numConnectAttempts = 0;
+            do {
+                try {
+                    Socket pushSocket = new Socket(m_server, m_port);
+                    OutputStream socketStream = pushSocket.getOutputStream();
+                    System.out.printf("Connected to VoltDB socket importer at: %s.\n", m_server + ":" + m_port);
+                    String s = String.valueOf(m_counter) + m_separator + System.currentTimeMillis() + "\n";
+                    socketStream.write(s.getBytes());
+                    pushSocket.close();
+                    m_counter++;
+                    return;
+                } catch (IOException e) {
+                    numConnectAttempts++;
+                    if (numConnectAttempts >= maxAttempts) {
+                        throw e;
+                    }
+                    try {
+                        Thread.sleep((int) (Math.random() * 1000) + 500);
+                    } catch (InterruptedException ignore) {
+                    }
+                }
+            } while (true);
+        }
     }
 
     abstract class DataPusher extends Thread {
@@ -379,8 +415,8 @@ public class TestImportSuite extends RegressionSuite {
         // Run data pushers until they are explicitly stopped.
         asyncPushDataToImporters(0, 3);
 
-        updateDeploymentFile(m_client, true, true, false);
-        updateDeploymentFile(m_client, true, false, false);
+        updateDeploymentFile(true, true);
+        updateDeploymentFile(true, false);
 
         waitForData();
     }
@@ -390,17 +426,17 @@ public class TestImportSuite extends RegressionSuite {
     public void testImporterDeploymentChanges() throws Exception {
         System.out.println("Removing and re-adding importers should cause them to restart");
 
-        SocketImporterConnector testConnector = new SocketImporterConnector(SERVER, SOCKET_IMPORTER_PORT, DELIMITER);
+        ImporterConnector testConnector = new ImporterConnector(SERVER, SOCKET_IMPORTER_PORT, DELIMITER);
         testConnector.tryPush(5);
 
-        updateDeploymentFile(m_client, false, false, false);
+        updateDeploymentFile(false, false);
         try {
             testConnector.tryPush(5);
             fail("Importer is still running even though it is no longer configured");
         } catch (IOException expected) {
         }
 
-        updateDeploymentFile(m_client, true, false, false);
+        updateDeploymentFile(true, false);
         testConnector.tryPush(5);
     }
 
@@ -409,7 +445,7 @@ public class TestImportSuite extends RegressionSuite {
     public void testAddRemoveImporterProcedure() throws Exception {
         System.out.println("Removing and adding procedures that the importers rely on should result in a restart");
 
-        SocketImporterConnector testConnector = new SocketImporterConnector(SERVER, SOCKET_IMPORTER_PORT, DELIMITER);
+        ImporterConnector testConnector = new ImporterConnector(SERVER, SOCKET_IMPORTER_PORT, DELIMITER);
         testConnector.tryPush(5);
 
         // importer uses CRUD procedure associated with this table.
@@ -421,26 +457,6 @@ public class TestImportSuite extends RegressionSuite {
         }
 
         applySchemaChange("CREATE TABLE IMPORTTABLE (PKEY bigint NOT NULL, A_INTEGER_VALUE bigint); PARTITION TABLE IMPORTTABLE ON COLUMN PKEY;");
-        testConnector.tryPush(5);
-    }
-
-    /** Verify that UAC does not start the importers while the cluster is paused.
-     */
-    public void testPausedUACWithImporter() throws Exception {
-        System.out.println("Catalog changes to a paused cluster must not restart the importers");
-
-        SocketImporterConnector testConnector = new SocketImporterConnector(SERVER, SOCKET_IMPORTER_PORT, DELIMITER);
-        testConnector.tryPush(5);
-
-        m_adminClient.callProcedure("@Pause");
-        updateDeploymentFile(m_adminClient, true, false, true);
-        try {
-            testConnector.tryPush(5);
-            fail("Importer is running even though cluster is paused");
-        } catch (IOException expected) {
-        }
-
-        m_adminClient.callProcedure("@Resume");
         testConnector.tryPush(5);
     }
 
@@ -478,25 +494,22 @@ public class TestImportSuite extends RegressionSuite {
     }
 
     /** Builds a CatalogContext for the import manager to use.
-     *
-     * @param clientToUse
      * @param includeImporters Whether or not to include the importers associated with this test.
      * @param unrelatedChange Whether or not to make a change that has no impact on the importer.
-     * @param extraneousImporterChange Whether or not to make a change that impacts the importer config but not the importer behavior.
      * @return New deployment file
      * @throws Exception upon error or test failure
      */
-    private static void updateDeploymentFile(Client clientToUse, boolean includeImporters, boolean unrelatedChange, boolean extraneousImporterChange) throws Exception {
-        VoltProjectBuilder projectBuilder = generateVoltProject(includeImporters, unrelatedChange, extraneousImporterChange);
+    private void updateDeploymentFile(boolean includeImporters, boolean unrelatedChange) throws Exception {
+        VoltProjectBuilder projectBuilder = generateVoltProject(includeImporters, unrelatedChange);
         File deploymentFilePath = new File(projectBuilder.compileDeploymentOnly(null, 1, 1, 0, 0));
         System.out.println("Deployment file " + (includeImporters ? "with" : "without") + " importers, " +
-                (unrelatedChange ? "with" : "without") + " unrelated deployment change written to " + deploymentFilePath.getCanonicalPath());
+                (unrelatedChange ? "with" : "without") + " command logs written to " + deploymentFilePath.getCanonicalPath());
         deploymentFilePath.deleteOnExit();
-        ClientResponse response = clientToUse.updateApplicationCatalog(null, deploymentFilePath);
+        ClientResponse response = m_client.updateApplicationCatalog(null, deploymentFilePath);
         assertEquals(ClientResponse.SUCCESS, response.getStatus());
     }
 
-    private static VoltProjectBuilder generateVoltProject(boolean includeImporters, boolean unrelatedChange, boolean extraneousImporterChange) throws Exception {
+    private static VoltProjectBuilder generateVoltProject(boolean includeImporters, boolean unrelatedChange) throws Exception {
         VoltProjectBuilder project = new VoltProjectBuilder();
         project.setUseDDLSchema(true);
         project.addSchema(TestSQLTypesSuite.class.getResource("sqltypessuite-import-ddl.sql"));
@@ -506,9 +519,6 @@ public class TestImportSuite extends RegressionSuite {
                 "port", Integer.toString(SOCKET_IMPORTER_PORT),
                 "decode", "true",
                 "procedure", "importTable.insert");
-        if (extraneousImporterChange){
-            props.setProperty("unrelatedChangeForTest", Boolean.toString(true));
-        }
         project.addImport(includeImporters, "custom", "tsv", "socketstream.jar", props);
         project.addPartitionInfo("importTable", "PKEY");
 
@@ -517,9 +527,6 @@ public class TestImportSuite extends RegressionSuite {
                 "port", "6060",
                 "procedure", "log_events.insert",
                 "log-event-table", "log_events");
-        if (extraneousImporterChange){
-            props.setProperty("unrelatedChangeForTest", Boolean.toString(true));
-        }
         project.addImport(includeImporters, "custom", null, "log4jsocketimporter.jar", props);
 
         project.setHeartbeatTimeoutSeconds(5 + (unrelatedChange ? 1 : 0));
@@ -538,7 +545,7 @@ public class TestImportSuite extends RegressionSuite {
         final MultiConfigSuiteBuilder builder =
             new MultiConfigSuiteBuilder(TestImportSuite.class);
 
-        VoltProjectBuilder project = generateVoltProject(true, false, false);
+        VoltProjectBuilder project = generateVoltProject(true, false);
 
         /*
          * compile the catalog all tests start with
